[update] test for App & Signin
