+++ src/App.test.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import App from './App';
 
-test('renders learn react link', () => {
-  const { getByText } = render(<App />);
-  const linkElement = getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+test('renders App element with className', () => {
+  const { container } = render(<App />);
+  expect(container.querySelector('main')).toHaveClass('App');
 });
